creating admin package and reservation package

+++ RTRT/src/main/java/com/project/RTRT/tables/TischRepository.java
@@ -2,5 +2,5 @@ package com.project.RTRT.tables;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TischRepository extends JpaRepository<Tisch,Integer> {
+public interface TischRepository extends JpaRepository<Tisch,Long> {
 }
